refactor: #2 リファクタリング対応
