+++ .vscode/settings.json
@@ -2,12 +2,5 @@
   "[python]": {
     "editor.defaultFormatter": "ms-python.black-formatter"
   },
-  "python.formatting.provider": "none",
-  "cSpell.words": [
-    "Dynmap",
-    "Kotoya",
-    "kzrm",
-    "Njgx",
-    "nohup"
-  ]
+  "cSpell.words": ["Dynmap", "Kotoya", "kzrm", "Njgx", "nohup"]
 }
